Upgrade to react-router 1.0.3
I also removed the two dependencies on events and when-node.  I'm not
sure why they were ever there in the first place, react-router has never
depended on either of those dependencies in its package.json or
bower.json files, and there's no mention of needing those in the
documentation, so I guess they were accidentally left in there when the
webjar was copied from another webjar that did depend on them?